+++ src/link.pyx
@@ -136,6 +136,16 @@ cdef class cLink(StructBase):
             self._value_error()
         return value
 
+    @property
+    def desired_capabilities(self):
+        cdef c_amqpvalue.AMQP_VALUE value
+        if c_link.link_get_desired_capabilities(self._c_value, &value) != 0:
+            self._value_error()
+        return value_factory(value)
+
+    cpdef do_work(self):
+        c_link.link_dowork(self._c_value)
+
     cpdef set_prefetch_count(self, stdint.uint32_t prefetch):
         if c_link.link_set_max_link_credit(self._c_value, prefetch) != 0:
             self._value_error("Unable to set link credit.")
@@ -143,6 +153,10 @@ cdef class cLink(StructBase):
     cpdef set_attach_properties(self, AMQPValue properties):
         if c_link.link_set_attach_properties(self._c_value, <c_amqp_definitions.fields>properties._c_value) != 0:
             self._value_error("Unable to set link attach properties.")
+
+    cpdef set_desired_capabilities(self, AMQPValue desired_capabilities):
+        if c_link.link_set_desired_capabilities(self._c_value, <c_amqpvalue.AMQP_VALUE>desired_capabilities._c_value) != 0:
+            self._value_error("Unable to set link desired capabilities.")
 
 
 #### Callback